# docs/swagger.yaml
definitions:
  model.CreateEmployeeInput:
    properties:
      hire_date:
        type: string
      person_id:
        type: integer
      salary:
        type: number
      title:
        type: string
    type: object
...
  model.CreateSubjectInput:
    properties:
      name:
        type: string
      specialty_id:
        type: integer
    type: object
...
  model.Employee:
    properties:
      address:
        type: string
      birth_date:
        type: string
      email:
        type: string
      employee_id:
        type: integer
      hire_date:
        type: string
      last_visit_at:
        type: string
      name:
        type: string
      notes:
        type: string
      password:
        type: string
      person_id:
        type: integer
      phone:
        type: string
      photo:
        items:
          type: integer
        type: array
      registered_at:
        type: string
      salary:
        type: number
      surname:
        type: string
      title:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  model.Subject:
    properties:
      faculty_id:
        type: string
      full_name:
        type: string
      name:
        type: string
      specialty_id:
        type: integer
      subject_id:
        type: integer
    type: object
...
  model.UpdateEmployeeInput:
    properties:
      hire_date:
        type: string
      salary:
        type: number
      title:
        type: string
    type: object
...
  model.UpdateSubjectInput:
    properties:
      name:
        type: string
      specialty_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for test.
  title: University System API
  version: "2.0"
paths:
  /admin/hub/employees:
    get:
      consumes:
      - application/json
      description: get employees
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/model.Employee'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get employees
      tags:
      - admin-employees
    post:
      consumes:
      - application/json
      description: create employee
      parameters:
      - description: Data for creating employee
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/model.CreateEmployeeInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create employee
      tags:
      - admin-employees
  /admin/hub/employees/{id}:
    delete:
      consumes:
      - application/json
      description: delete employee
      parameters:
      - description: ID for deleting employee
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete employee
      tags:
      - admin-employees
    get:
      consumes:
      - application/json
      description: get employee by id
      parameters:
      - description: ID for getting employee
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get employee
      tags:
      - admin-employees
    patch:
      consumes:
      - application/json
      description: update employee
      parameters:
      - description: ID for updating employee
        in: path
        name: id
        required: true
        type: integer
      - description: New information for update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateEmployeeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update employee
      tags:
      - admin-employees
...
  /admin/hub/subjects:
    get:
      consumes:
      - application/json
      description: get subjects
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/model.Subject'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get subjects
      tags:
      - admin-subjects
    post:
      consumes:
      - application/json
      description: create subject
      parameters:
      - description: Data for creating subject
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubjectInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create subject
      tags:
      - admin-subjects
  /admin/hub/subjects/{id}:
    delete:
      consumes:
      - application/json
      description: delete subject
      parameters:
      - description: ID for deleting subject
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete subject
      tags:
      - admin-subjects
    get:
      consumes:
      - application/json
      description: get subject by id
      parameters:
      - description: ID for getting subject
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Subject'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get subject
      tags:
      - admin-subjects
    patch:
      consumes:
      - application/json
      description: update subject
      parameters:
      - description: ID for updating subject
        in: path
        name: id
        required: true
        type: integer
      - description: New information for update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubjectInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update subject
      tags:
      - admin-subjects
...
swagger: "2.0"